bug fixes from Overhaul: - Data type changes - Templates removed
git-svn-id: https://ja2svn.mooo.com/source/ja2/trunk/GameSource/ja2_v1.13/Build@871 3b4a5df2-a311-0410-b5c6-a8a6f20db521

+++ Tactical/UI Cursors.cpp
@@ -570,7 +570,7 @@ UINT8 HandleActivatedTargetCursor( SOLDIERTYPE *pSoldier, UINT16 usMapPos, BOOLE
 				{
 					//AXP 29.03.2007: Rooftop CtH fix. See below.
 					INT8 bTempTargetLevel = pSoldier->bTargetLevel;
-					pSoldier->bTargetLevel = gsInterfaceLevel;
+					pSoldier->bTargetLevel = (INT8) gsInterfaceLevel;
 
 					UINT32 uiHitChance;
 					uiHitChance = CalcChanceToHitGun( pSoldier, usMapPos, 0, pSoldier->bAimShotLocation );
@@ -590,7 +590,7 @@ UINT8 HandleActivatedTargetCursor( SOLDIERTYPE *pSoldier, UINT16 usMapPos, BOOLE
 				//AXP 29.03.2007: Rooftop CtH fix. Temporarily set soldier targetlevel to the real interface
 				//level (selected with TAB), so the first shot after switching TL is calculated correctly
 				INT8 bTempTargetLevel = pSoldier->bTargetLevel;
-				pSoldier->bTargetLevel = gsInterfaceLevel;
+				pSoldier->bTargetLevel = (INT8) gsInterfaceLevel;
 
 				UINT32 uiHitChance;
 				uiHitChance = CalcChanceToHitGun( pSoldier, usMapPos, (INT8)(pSoldier->bShownAimTime ), pSoldier->bAimShotLocation );
